refactor pm quality review action lookup routes
